+++ src/deploy/devnet/registry_main.ts
@@ -1,2 +1,2 @@
-import { Registry } from '../../Registry';
+import { Registry } from '../../doot/Registry';
 import { MultiPackedStringFactory } from 'o1js-pack';
