+++ src/doot/Registry.test.ts
