+++ src/doot/Registry.ts
