+++ src/doot/barter/Barter.sol
